V1.2.2
Added specificline keyword argument to get_code function in /src/loader.py/.
Worked more on the RegexString datatype.
Modified the call_error function to show more information, now including the line number of the current file and the full path of the current file.
The GOTO statement now does not require a colon or any right hand arguments, only a line number argument. The preferred usage now is to use a shorthand condition to specify which line the line cursor moves to. As well as adding negative line numbers, you can now also have a null line number, or a line number of 0. Leaving the line number as 0 or null will keep the cursor still, which works well with shorthand conditions.