V1.2.4
Did a lot of major changes in this update, including optimisation fixes, improvements in code readability, tokenisation fixes, and the addition of some new major features.

These new major features include a new function declaration and calling system, as shown in the files within /tests.
A new system for yielding has been added, to stop functions from printing unwanted text or returning unwanted values.
The datatypes have been improved more, and next some new datatypes will need to be added, such as Arrays.

This version of the code is a stable Release compared to the previous ones, and a Windows build will be added to the /dist folder in about 12-14 hours from the current time of this commit.