V1.3.0
This was a major update.
Functions have been improved, IF-ELIF-ELSE chains have been added, and more.
Procedures have been fully removed.
When calling a function, the function's return value is automatically assigned to the global variable '_'.
Booleans return their value in uppercase when converted to a String.
Code readability has been improved in some areas.
Error checking has been improved again, bugs have been fixed in some areas.
EXIT function now takes no arguments.

GOTO statements now only work in areas that are not protected by scope definers, or '{' and '}'.

IF statements were probably the biggest part of this update, and also the reason why the version number increased from 1.2.4 straight to 1.3.0.

This version has been exported to the /dist/ folder.
This version is also a stable version compared to the first few commits, the only things to note would be the debugging that gets logged when an expression error occurs, but expression errors should only occur due to errors that I have caused myself accidentally, so if debugging shows up then it should be reported as an error.