1.3.2
Lots of bug fixes.
Added HAS keyword, currently only works for String.
Arrays can now be initialised from Null.
Fixed a lot of Array bugs.
The Null object's value is now None, instead of 'null'. Before, BOOLEAN NULL would evaluate to TRUE, because Null's value was a true string, 'null'.
Error checker only prints current line if not inside a method.
Fixed consistency issues with the evaluate function's calls.
Removed revertkeys as it is now redundant.
Added AND and OR keywords.
Made the yielding system more useful. When yielding only one value, that value is returned, when yielding no values, Null is returned, when yielding more than one value, an Array is returned.

I wrote a short program in maintest.mdcl to print out the instructions for getting to the current file via command line.
I also added a JOIN function which will be useful for development and may also be added into the built-in functions list in the future.