1.3.3
Another major update, FOR loops have been added, the TO and SKIP statements have been added, and the ECHO built-in function has been added.
When iterating over an Array with a FOR loop, the current iteration item will always be stored in `,` (comma). To retrieve this value, just type a comma into your code as if it were a letter, and you will get the current iteration item, or if you are outside of a FOR loop, the last iteration item of the previous FOR loop.

Array datatype now has a SUB function.
Improvements to Array initialisation.

README was updated to include new information and better formatting.

This version has been exported to the /dist/ folder.