1.3.4 (Complete)
This is the first part of 1.3.4, it is a stable update but some more work is required to fully release it.
Try-catch statements are fully functional (only to the extent of the tests that I have tried), and the SIG and RAISE statements have been added. SIG INT 0 will disable keyboard interrupt, SIG INT 1 will re-enable it. SIG also takes other signal types, which currently include INT, TSTP, and SEGV. The 0 or 1 after SIG can also be replaced with FALSE or TRUE, or even just a plain expression, as it will be evaluated and converted to a Boolean whether you specify a Boolean type or not.
KeyboardInterrupt and EOFError is now properly caught, as well as RecursionError and also Segmentation Faults and other fatal errors.
A few new error types have been added, and I'm planning on making a statement for adding new error types inside MDCL for more personalised uses.

Removed the GOTO statement, and majorly changed the run function
Also added the perform_try_catch function.

Finally, I'm going to rewrite the run function in the next update to match the syntax of the RAISE statement, where the line gets evaluated BEFORE the function picks out values from the line.