1.3.4 (Fully completed)
This is a full release of 1.3.4.
Lots of bugs have been fixed, and this is now a stable release.
I have written a small data interchange language using MDCL in the /tests folder.
This language is currently very flexible, and can be shortened to extremely small sizes, and is on average 20% smaller in size than the equivalent written in Python.

This version has been exported into the /dist folder.