1.3.6
A completed release of 1.3.6.
This version has been exported to the /dist folder.

Tokenisation improvements, including the use of regex.
Improvements to datatypes.
The last value to be assigned to a variable is now stored in the global variable '='.
For loop improvements.
I'm experimenting with a for loop which iterates over every item in a list at once using threading, and this update may be released in the near future.
MDCL now catches ZeroDivisionErrors and deals with them accordingly, raising the 'zerodivision' error.

Slight improvements to the way that datatypes are converted from Python types to MDCL types.
Reserved names can now not be ovewritten in any way, and you can not assign a value to a reserved name.
The WAIT built-in function has been added.