1.3.6.2
Removed redundant code.
IMPORT statement has gone back to it's previous version.
You can no longer overwrite built-in errortags.
Prefixing a function call or value with an underscore ('_') will not echo it's value to the console. There is an error with this new addition, it will need to be fixed in future updates. The underscore is now completely omitted in any place or usage, which I didn't realise until just now.
`readonly` errortag added.