1.3.8 (Incomplete 1)
I am currently working on the ability to import functions from Python modules for use within MDCL.
It is not complete yet, but at the moment you can import most variables and functions into your MDCL file.
The only issue that needs to be fixed is that you can't run any of the functions that are converted and imported.