1.3.9
Added DATE datatype.
A DATE is the equivalent to Python's datetime.datetime object. It can be initialised with 0 or 1 arguments. With 0 arguments, it's value becomes the current time.
The D string will be added in a future update to initalise DATE datatypes through a different method.
Added Debug.ts method for printing types of values in a list.

This version has been exported to the /dist folder and is a stable update.

Next update will be quite a large one, as customisable datatypes will be added.