1.4.0
You can now make new datatypes!

This is a fairly large update, as the DATATYPE keyword has been added, along with many other features.

The Datatype class has been renamed to BaseDatatype, and it's do_action method has been modified. The do_action method can now detect and handle an action when the action it retrieves is of type Function or BuiltinFunction.
The String built-in datatype can now handle all built-in datatypes in the language as well as other datatypes you make throughout your code. It is used to convert anything into a printable String.
You can now define Special Methods with the NEW keyword, and these are currently used with the DATATYPE keyword to define functions for use within a new Datatype. Special Methods are prefixed with '!' and can be named ECHO, or any of the other valid keywords that can be used between Datatypes.
When defining a Special Method, you may list anything as valid types for the Special Method's arguments, so as to make it easier to use when a Datatype may not exist yet.
The PYIMPORT keyword has been fully removed, but still remains commented in the code.
Function make_evaluable has been created for recursively turning a list of Datatypes into a list of evaluable items which can in turn be printed.
kwds variable in evaluate function has been moved to global scope and renamed to mdcl_keywords.
In a future update, the '.' character may be used to get attributes from Datatypes, for example the Datatype's value.
Fixed issue with the '_' global variable not being acknowledged.
Added error tag 'defined' as 'ALREADY DEFINED VALUE'.
Python's datetime.datetime converts to MDCL's Date and Python's datetime.timedelta converts to MDCL's Timer.
Re-arranged some variables to fix issues with the String datatype not being able to get the current list of datatypes.

This version has been exported to the /dist folder and is a stable version.