1.4.1 (Incomplete)
Removed Timer datatype.
Added Timedelta datatype.
Timedelta is the equivalent to Python's datetime.timedelta.
Added the ADD, SUB, MULT and DIV keywords for Timedelta.
Added the ADD and SUB keywords for Date.
Started writing the Slice datatype.
The Slice datatype will be initialised with a slice surrounded by '(' and ')'.
A slice will be used to get item/s from an Array, String, or any other datatype that has a SLICE method.