1.4.1
Comments are a thing now!
Block comments are made by surrounding code or text with /* and then */.

Slices are now fully functional and very useful. You can use a Slice to chop up an Array or String to be the way you want it to be. Slices in MDCL are just like Slices in Python, and they almost look the same too.
A Slice in MDCL is surrounded by brackets, '(' and ')', and has three arguments within itself separated by colons. When initialising a Slice via this method, you can omit any amount of these colons, as sometimes they aren't needed. Check /tests/newtest.mdcl to see examples of Slices being used.

Next update, Functions will be called using the new operator '!', and Functions will act as regular variables rather than being kept in a separate area.

This version has been exported to the /dist folder and is a stable version.