1.4.2
Major changes to the way Functions and BuiltinFunctions work in MDCL.
A Function or BuiltinFunction can now be initialised with another Function or BuiltinFunction object as the first argument, rather than assigning one using name, args, and code.
Functions can now be displayed, and won't always run when you type their name somewhere.
To run a Function, you now have to prefix it's name with '!', and this operator can also be prefixed before other types if they are close enough to a Function to be called, which at the moment is impossible or almost impossible to do through MDCL code, so only Functions and BuiltinFunctions use the '!' operator.
Removed the functions dictionary.
All BuiltinFunctions and Functions go into local_vars now.
The eval_functions method is now reduntant.

This version has been exported to the /dist folder and is a stable version.