1.5.2
This is another fairly large update. A lot of the memory system has been improved, but it still lacks efficiency and keys are constantly being repeated. The temporary fix for this is to use the CLEAN function, which completely wipes out every unused chunk of memory, reducing the memory load by about 80% percent.
This will gradually be fixed properly in future updates.

This version is partly stable, with some known issues.
This version has been exported to the /dist folder.