1.5.1
Added Argtypes Datatype.
Argtypes are made the same way as Slices, but with ! at the start.

Example:
(! : String)

The above creates an Argtypes object that takes String as it's first argument.
Argtypes are used with the new InlineFunctions to create Functions without having to specify a Function name.

InlineFunctions can be created like this:
() => {
    /* CODE HERE */
};

This creates an InlineFunction that takes no arguments, and outputs nothing.
Printing an InlineFunction returns <InlineFunction>.
InlineFunctions are now going to be used instead of the NEW keyword, and the NEW keyword is now going to be used instead of the DATATYPE keyword to make Datatype creation easier, and to move more toward the fact that Datatypes will eventually become Objects, rather than just Datatypes.

This version has been exported.