1.5.4 - Quick Var Assignment
Updated tokeniser, added 4 new tokens:
+:
-:
*:
/:

These tokens are for quick variable assignment.
You can now do Python's equivalent to +=, -=, etc with +:, -:, etc.
Doing this:
a +: 1;
Is the same as this:
a : a + 1;
Or this:
a : a ADD 1;

Using these new tokens, you can quickly modify variables with more ease than having to reassign to them the old way. Of course, you can still use it the old way if you need to for any circumstance.

Fixed a few bugs, assigning a variable to nothing now assigns it to NULL.
In the past, doing this:
a:;
Would fail to make a equal to NULL, but this has been fixed.

This version has been exported.