1.5.6 - [] and () switch
The operators [] and () have now switched places. Yep.
Arglists and Slices are created with [] rather than (), and Arrays and groups are made with () instead of [].

Fixed up the CALL methods for all the Built-in Datatypes. Prefixing a value with ! will now return the NOT of it's value, which will usually be FALSE if the value is TRUE, and TRUE if the value is FALSE.

Fixed some Arglist issues.
Renamed eval_datatypes to pre_evaluate, because it now evaluates the getattr operator as well as datatypes. The getattr operator is '.', it's used for getting attributes from values.
Added GETDATA BuiltinFunction, it is used to get the data object from a Datatype. It is currently returned as a NULL Datatype, but it's value is a string formatted version of the data object of the Datatype that GETDATA was called with.

This version has been exported.