1.5.8 - Removing the silly ':' operator
In this update, the ':' has been removed from if statements, while loops, for loops, and try catch statements.
This means you can write them all much more easily, and save yourself some time and effort by leaving them out altogether.
The statements all look a lot more appealing now that the ':' operator has been removed.

Fixed bug where Booleans return in String form as TRUE or FALSE instead of true or false.
Fixed Slice bug where [:-1] or any number instead of -1 would fail to run because of a missing condition check in the do_slice method.

Added Signal 'ECHO', which is by default true. If this signal is true at any point during code execution and a value is given directly to the interpreter, it will echo it to the console, just like what happened in previous versions.
However, disabling this signal will result in the console being completely silent unless using the echo function, which is what happens with most languages anyway.

Slight performance increase in for loops.
Major performance increase in evaluate function. The old line `new = [a for a in exp]` was actually extremely slow compared to the new line `new = exp[:]`. The speed now is about 1.9 to 2 times the speed it was before, which will hopefully greatly increase the overall execution speed of the interpreter.

The split_list function now has better options, including the ability to have a lambda function or regular function as it's split_at argument rather than a string.
Passing a lambda function or regular function as it's split_at argument will split the list as usual, except it will run the function with the current list item as it's argument, which can help a lot with extra list splitting customisability.

This version has been exported.