1.5.9 - Builtin datatype functions!
All datatypes (except Functions) now have a set of builtin functions that you can use!
Currently all datatypes have the map function, and Strings have a large list of different string-manipulating Functions which you can see in the list of changes.

Made some changes to the Boolean datatype so that it's value is now a Python bool, rather than an MDCL Integer.
Booleans print their values in lowercase now, which should have been happening in the last few updates.

Fixed a few bugs relating to the recent mass rename.

One of the other big features in this update is the fact that the elif and else keywords are now optional!
This means that you only need to start an if statement with the if keyword, then you can either add code blocks prefixed with conditions for elif statements, or just add an empty code block straight after the previous one to act as an else statement. Of course, if anyone prefers to use the elif and else statements in if-elif-else chains, the functionality is still there, and there is no difference whether you use them or not.

Reading from files now removes carriage returns from the text before returning it as a String to fix some issues with file reading.

The echo function is now used to print raw values, and doesn't take sep or end values anymore. The old echo function has been moved and renamed to write.

In the following update, 1.6.0, I have a plan to change function calling once again, but this time it will almost definitely be a permanent change.
I'm planning to have functions called with the name of the function, plus an Array on it's right hand side full of arguments. It'll look something like:

myFunction(1, 2, 3);

Which looks like most other language function calling systems, so it shouldn't cause any issues.

This version has been exported.