1.6.0 - Function calling overhaul
This update adds exactly what the title says, plus A LOT more.

The biggest change here is the fact that functions are now called with a function name plus an Array on it's right hand side, like this:

function('a' 'b' 'c');

In the following updates I'm hoping to also remake the Array creation system, most likely so that commas must be included to define an Array.

Removed a few lingering unused imports.
Renamed some error messages.

Added exec and eval MDCL functions.

You'll have to see the rest for yourself :)

This version has been exported.