1.6.1 - Array commas and CLI Interpreter
Arrays are now defined with values separated by commas, and placing more than one value in a row will cause an error.

When running the mdcl command with no arguments, you now get a nice CLI Interpreter where you can enter MDCL code just as you would in a text file, but your code gets interpreted immediately and returned to you in the console.

for loop temporary variable is now '#' instead of ',' to avoid Array comma issues. This will change at some point in the future.

Surrounding an expression with backticks (`) is now the equivalent to surrounding an expression with regular brackets in previous versions of MDCL.

Removed replaceargs function as it was redundant.
Removed eval statements to accommodate the new use for backticks.
Made some minor changes to the exit function.

This version has been exported.