1.6.3 - Math, man
This is the smallest update I've released in quite a while.

Fixed a bug with parentheses turning everything into Arrays whether they wanted to be turned into Arrays or not.
Parentheses act how they should now, evaluating the code within them before the rest of the code in an expression.

Placing multiple tokens in a row will multiply them together, which is what happens in algebra. This would most commonly be used with expression like 2x + 1, where 2x would evaluate to 2 times x, whatever x's value may be.

This version has been exported.