1.6.7 - Efficiency Improvements and Underscores
Fixed A LOT of bugs.

Underscores are now valid next to words, so you can name variables with underscores in their names.

Fixed a bunch of naming issues.
Added call_function function for calling functions. Very useful.
When you run an MDCL script, the CWD no longer changes to the script's directory, it stays as the directory from which the script was executed, which makes sense.

The greater than and less than signs now actually check whether things are greater than or less than each other.
Added a shorthand if-else statement which looks a lot like JavaScript's shorthand if-else statement, but not EXACTLY like JavaScript's if-else statement because that wouldn't be as fun. So it's basically the exact same except with a backslash instead of a colon, and you don't even really need the whole second half because the second half defaults to null if the condition is false.

It looks like this:

condition ? "Yeah, it's true." \ "Nope, it's not true."

Obviously condition is the condition to check, and the first string is what is returned if condition is true, and the second string is what is returned if the condition is false. If the second half (including the backslash) is omitted, then it's the same as having backslash null (\ null) at the end.

Added MDCL functions tokenise and tokeniseFile for if you ever wanted to tokenise an MDCL script from within an MDCL script. Useful, I know. ;)

Completely remade the if-elif-else chain so that they're now all separated into their own fancy keywords, and they work using a global Python variable called previous_condition, which is a boolean that says whether the previous condition in any if-elif-else chain was true or false, which can be used in the next elif or else statement to decide whether it should execute or not. Which is handy because now you can make your code look all fancy, and I can also add features later for else statements to be placed after all kinds of other keywords, like for, while, and whatever else I can think of.

This version has been exported. You know what, they're all getting exported, so I'm just gonna say when a version is NOT exported.
So assume that they're all exported unless I say otherwise.