1.6.8
This version will not be exported.
This may be one of the last commits of MDCL in this GitHub Repository, as I will slowly move towards MDCL becoming a compiler.
This update is mainly just the stepping stones that can be used with the MDCL compiler in the future.
Of course this update can still be used as an interpreter, and this won't change, at least in this Repository.